Adds (some of the) thumbnails for collections, with tests
To check the validity of the the thumbnail URLs, run

  RSPEC_CHECK_LINKS=true rspec spec/features/collections_spec.rb

I removed all of the sample images, e.g. "clip_1.jpg" so we can see what failing
tests look like.

Not convinced this is the best approach at testing CMLess content. It might be
better create a validation module that could be unit-testable against mock
objects. I am, however, convinced that there should be some kind of testing on
CMLess content, since putting bad stuff in could break the site.